hardqode/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
